Rename Data Availability Layer Client
