+++ da/registry/registry.go
@@ -2,14 +2,14 @@ package registry
 
 import (
 	"github.com/celestiaorg/optimint/da"
-	"github.com/celestiaorg/optimint/da/lazyledger"
+	"github.com/celestiaorg/optimint/da/celestia"
 	"github.com/celestiaorg/optimint/da/mock"
 )
 
 // this is a central registry for all Data Availability Layer Clients
 var clients = map[string]func() da.DataAvailabilityLayerClient{
-	"mock":       func() da.DataAvailabilityLayerClient { return &mock.MockDataAvailabilityLayerClient{} },
-	"lazyledger": func() da.DataAvailabilityLayerClient { return &lazyledger.LazyLedger{} },
+	"mock":     func() da.DataAvailabilityLayerClient { return &mock.MockDataAvailabilityLayerClient{} },
+	"celestia": func() da.DataAvailabilityLayerClient { return &celestia.Celestia{} },
 }
 
 // GetClient returns client identified by name.
